Fixed shell script for making mpi files
The zfix shell script had only one name for a temp file, "/tmp/zxx".  This did
not allow two builds on the same machine to run the script at the same time.
This is a possible cause of build failures, such as the corrupted mpi.h seen
today in https://jenkins02.trafodion.org/job/build-product-release/141
which had this error:

... /sqf/export/include/mpi.h:837:1: error: unterminated #ifndef	##(MPI)

The temp file name is now "/tmp/zxx.$$" which adds a pid to the end of the name.

Change-Id: I20e6549342fe792fba30524cbc5fe14cca833a1d