Package log4cpp source into core product tgz file
The log4cpp source files used to build trafodion/core have small differences
from files downloaded from http://sourceforge.net/projects/log4cpp version
1.1.1, so they should be provided to the end user.

This change packages the log4cpp directory of https://github.com/trafodion/core
into a tar file, and puts that in the trafodion core product's .tgz file as:

	sources/log4cpp-1.1.1.tar

When Trafodion is installed, this file is available to the end user.

This new file has tar format, not gzip format, and hence its name has .tar
instead of .tgz.  Files with the tgz extension are excluded from packaging
into the core product because "*tgz" is in sqf/build-scripts/package.exclude.

In sqf/Makefile, 4 obsolete lines were deleted, because this build does not
compare the package contents with a bom (bill-of-materials) list.

The packaged source files include the license used by log4cpp, in the file
named COPYING which starts with:

                      GNU LESSER GENERAL PUBLIC LICENSE
                           Version 2.1, February 1999

     Copyright (C) 1991, 1999 Free Software Foundation, Inc.
         59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
     Everyone is permitted to copy and distribute verbatim copies
     of this license document, but changing it is not allowed.

In the second patch set, added the new tar file to two .gitignore files
so builds in Jenkins do not fail the test for untracked files in the
check-git-status.sh script.

Change-Id: I8db0559128e3f57653ad3cf08f2b553b225a550c