Corrected typos in user messages
Corrected about 15 typos in messages which can be reported to the user,
for example, "loaction" which should be "location".

While changing files, also removed unnecessary spaces from the end of
source lines, error messages and strings.

In backup_restore_functions.sh, did not change 9 uses of "hsdf" to "hdfs" in
shell coding, for example, line 212:    local hsdf_loc="$1"
While this is user-visible, it is not likely to be seen, and perhaps this
is used with other non-Trafodion scripts which already use the better
spelling, $hdfs_loc.

Change-Id: I311da5560057d441cb6ab80e91bef6d0a4f6ae3c