Invoking sqenv.sh repeatedly does not change shell environment
Bug 1401216 reported that sqenv.sh is not idempotent.  Previously this
was fixed by using shell function remove_duplicates_in_path() for the
variables that grew.  However, MANPATH was missed.  This change adds
one line which ensures MANPATH does not grow.

Trailing spaces which were added to sqenvcom.sh in the last few months
were deleted in this change.

Change-Id: Id0bca1cfbf79d0f215667cbbf205b1b5180b6e78