Fix 1455585: Error 1234 returned trying to run SQL
If code is refreshed with https://review.trafodion.org/1635 (which adds the
first phase of column level privileges) authorization is enabled, and running
in debug mode; error 1234 is returned and you are not able to drop or
initialize authorization. SQL requests fail.

This delivery fixes the issue.  First it allow operations to proceed even if
there are missing privilege manager tables.  Second, we have turned off checks
for column level privileges (can be turned on by setting cqd CAT_TEST_BOOL).
Lastly, we do not run the column privileges regression test.

Change-Id: Ic427c4e9413b6c7208313c0e0755ca6aabd8a2cd