Fix 1450193: User ... plans generated without using histogram stats
The issue occurs because the person compiling the query does not have select
privilege on the HISTOGRAMS table so statistics cannot be retrieved.
We support a special interface to allow metadata tables to be accessed without
privs for operations performed during compilations. However, this interface is
not being used for HISTOGRAMS tables during query compilation. The change is to
use this interface when gathering statistics during a query compilation.

Added exception handler around setting/resetting the parserflag.

Change-Id: I13f9a4bd870b3978e29aa1dad251c24ed400f8ff