Part 2 of ALTER TABLE/INDEX ALTER HBASE_OPTIONS support
blueprint alter-table-hbase-options

This set of changes includes non-transactional HBase semantics
for ALTER TABLE/INDEX ALTER HBASE_OPTIONS. As part of this
change, two sets of logic that manipulate HBase options in
the CREATE TABLE code path were refactored so that logic
could be reused by ALTER TABLE/INDEX. In both these cases,
the logic in question was taken out of a create-side method
and moved into a separate method.

Still to come is the equivalent transactional support, and
also a regression test script that exercises this feature.

With this change, ALTER TABLE/INDEX ALTER HBASE_OPTIONS works
but only in a non-transactional manner. That is, if the
operation is aborted after the request has been issued to HBase,
the change to the Trafodion metadata may be rolled back but
the change to the underlying HBase object may still complete.

Change-Id: Ied11d0b2ac7b7d65110d39a1e5a7fbbac1e1f827