Additional fix for 1452424 VSBB scan cause query to return wrong result
VSBB scan was returning wrong results randomly  even for the same query.
An extra null byte was added in the row id at times while setting up the
unique key. This caused a shift in row id parsing at java layer leading
to row not found for all row ids after this extra byte.

In addition, VSSB select was not getting GET_EOD or GET_NOMORE queue
entries at the end of the query. Hence, the rowset was never getting
closed causing a resource leak.

Also, retained the other changes that helped in debugging this issue.

Change-Id: I6a807bddc8edaff2f4140931d4f228e94badcc05