Reworked fix for 1452424 VSBB scan cause query to return wrong result
The assumption that rowID should be exactly equal to the calculated
key length is not correct. Trafodion SQL engine uses a null extra byte
to ensure that the row ID is not found in case of data conversion errors
Hence direct ROWID buffer format is changed to accommodate this.

Now the format is
numRowIds + rowIDSuffix + rowId + rowIDSuffix + rowId + …
rowIDSuffix is '0' then the subsequent rowID is of length
                            =  passed rowID len
               '1' then the subsequent rowID is of length
                           = passed rowID len+1

Change-Id: I07a283895f6f9c652b3f933bcf0330b69ee2d300