Repository support in MT server #2
This is the second installment of code for the repository functionality
in the DCS multi-threaded server/T2. Adding query statistics code.

Note:
This code is for compile only for now and not part of any active code path.
Most of the this is existing code from MXOSRVR side and is made to fit
into T2.

There are some temporary to-do comments in the code which shall be
removed eventually.

Change-Id: Iff4a3ec9d87ad6d3ce7c2e1882126368ce9e6d6b