Copyright corrections
Put correct form of Apache License into many source files.

The tdm_odbcSrvrMsg.mc file is used on Windows, and is
treated as binary on Linux.  However, it is visible in
Linux with "cat", and editable with "notepad" on Windows.

In patch set 2, updated sql/regress/newregr by deleting some
unused files, updating copyrights in scripts, and removing
copyrights from front of expected test results.  This is not
a final list of the changes needed, since the test suites in
the "newregr" directory need more work before they run correctly.

Change-Id: If0ea5331b822764191c839c38dfd84218612fe7a