fixes bug 1464364
Update Stats on a table in a schema with a delimited identifier name
failed with a funny syntax error. Turns out in a new sqlci session,
Showddl on the same table fails the same way.

The root cause was the schema name was passed to a ComObjectName
constructor in internal format, when external format was required.
This has been fixed.

Change-Id: Ic81ba3d69e006264340fbdcbff6fb455593e30cf