Added a hive test to core
Hive tests fail at times since there is no test that is part of the core
suite and sometimes people miss running them. So added one test so it's
part of the gate tests. It's a very short test that runs in a few
seconds. The Hive suite should be runfor more comprehensive coverage.

Change-Id: I75862841e52a71583632408fd06f6cc5fd6d83ac