Privilege manager fixes for 1438896 and 1465356
1438896: Internal error during create or replace view

The objectUID check when getting privilege information was not correct.

1465356: Revoke privilege from role returns dependent object error

There is a check in mainline revoke code to determine if the object type is a
view and if the SELECT privilege is no longer applicable. If so, then the
dependent error is returned. However, this code is incorrect and actually the
correct code exists in the gatherViewPrivileges method. The view check has been
removed.