Prepare "ROLLBACK" and "COMMIT WORK" at connection phase
Get "ROLLBACK WORK" and "COMMIT WORK" transaction prepared
during the connection phase. This is for Rowsets functionality
purpose, either the "COMMIT WORK" Transactioni need to be executed
to make it work, or "ROLLBACK WORK" shall be executed for failure.

PS: This includes a few interface changes related to Rowsets logic,
the rowset implementation code will be committed in another one or
more commits in the near future. This is step 1 check-in.