Fix for [TRAFODION-14]
upsert or merge into a table with indexes can result
in inconsistency between index and table

This change avoids the HBase issue of rows getting lost when the
 delete and put operations are done in the same milliseconds
timestamp dimension for the same rowid.

In the trafodion_delete operation execution, the row is skipped
for deletion if the new and old rowid evaluates to the same value.

This needs change from the compiler sub-task in the JIRA to complete
for it to be effective.

In addition removed yet another memory allocation for the delete
operation.