[[Trafodion 1133]] [[Trafodion 1088]] [[Trafodion 1481]]
Trafodion 1133 and Trafodion 1088 allow installer to add a node.

Trafodion 1481 seperates rpm installs into its own script so that it can be
run during an upgrade (to make sure all new packages needed are added). Bug still needs to be fixed is that it runs too many times (is ran once per node) and doesn't need to be. Will create a new jira for that small bug. Would like to check this in to prevent issue when upgrading.