Changes to sqenvcom.sh file to make it easier to run on Apache HBase without one of the distros.
The script now checks for common HBase and Hadoop setup scenarios, for
example with environment variables HADOOP_PREFIX, HBASE_HOME and
HIVE_HOME set.

This work is not complete yet, we need to document the entire process
and probably need to make a few small code changes to fix interfaces
that went away or are not yet added in Apache Hadoop/HBase.