AFODION-1492] Allow specification of MaxHeapSize for JVM started by UDR server
This is a fix for the second of the two issues described in this JIRA. The first
has been addressed by narendra in a separate checkin.

In this checkin we allow multiple methods to specify JVM startup options for the
UDR server process. There are three methods applied in this order with the last
specification for each property superceeding previous specifications.
1. A default default of 512MB for UDR JVM's max heap size
2. Specify any desired property including -Xmx in a new configuration file
3. Specify any desired property using TRAF_UDR_JAVA_OPTIONS env variable

The new configuration file is at $MY_SQROT/conf/trafodion.udr.config. It has
detailed comments on how properties should be specified in the file. The location
(including name) of the configuration file can be overridden with the env variable
TRAFUDRCFG.

The TRAF_UDR_JAVA_OPTIONS env variable must be specified in $MY_SQROOT/etc/ms.env
for it to be visible to the UDR server. Multiple properties can be separated by a
'\t' (TAB) chracater or the '\n' (newline) character