[TRAFODION-1516] An trafodion infrastructure program mxssmp dumps core at times
ssmp was dumping core sometimes when the DCS MASTER tries to cancel a query using pid.
When the pid is more than 32767,  the executor was not handling it correctly resulting
in negative number for pid. This resulted in partial junk message being sent to SSMP.
The junk message causes ssmp to dump core sometimes.