[TRAFODION-1575] Avoid transforming update into delete and insert
commands to improve performance of update statements. When a table has
check constraint the update command is transformed into insert and delete.

Added code to evaluate constraint expressions in all TCBs so that
constraint checking is done correctly. But, the subtask [TRAFODION-1610]
needs to implemented before this change is exercised.