[TRAFODION-1423] indexes are now created in aligned format by default
The indexes row format is now independent of the table. The table
can be in format that aids in pushing down the predicates to
hbase, but the index is now created in aligned format by default.

SHOWDDL <table_name> now shows the row format of the index when it
is in aligned row format.