Added Google Analytics Tracking to all .html pages. Code: UA-72491210-1.
For the website pages, this is accomplished by adding
<googleAnalyticsAccountID> tag to the site.xml file.

For asciidoc files, an extension from asciidoc-extension-labs provides the
needed function:
https://github.com/asciidoctor/asciidoctor-extensions-lab/blob/master/lib/google-analytics-postprocessor.rb

*** LICENSE NOTICE ***

Like asciidoc itself, this extension uses the MIT license.

In addition, it has a dependency on jruby.

*** RELEASE INSTRUCTIONS ***

Please build the website and the asciidoc files as documented at:
http://trafodion.apache.org/document.html#publish

There no real need to replace the .pdf files but it's simpler to replace
all the latest and 1.3.0 documentation.