[TRaFODION-1699] Updated web site to rely on Contributor Guide on the wiki. Changed download to point to real binaries hosted on the build-result site. Ran spell checks on all the guides and fixed formatting errors as I saw then.
The *-redirect.md pages redirects Contribution and Team pages to the wiki
rather than doing a direct link. This is to allow us to track access using
Google Analytics, which is not possible with direct links to the wiki.

Also tried out Steve's new build for the web site, which works very well.
Thanks!

Please publish the documents to 2.0.0 and update the web site. The deleted
.md file can be deleted from the web site's SVN repository, too.