Addresses executor TEST131 failure - TRAFODION-1796
Executor TEST131 is failing due to a testware issue on how default
schemas are determined.  This delivery fixes the failure by creating
a special schema that is created and dropped by the test.

This does not resolve the problem written up in TRAFODION-1796 but it
allows the test to run correctly in the SQL test environment.