[TRAFODION-1614] Improve time to build external dependencies
This script is for one-time set-up of build environment (aka TOOLSDIR),
but an hour or more is excessive. So to shrink the time, we kick off all
the builds in background.  All the builds are independent, except
that LLVM depends on udis86.

LLVM is the long pole in build time. Biggest gain is building the
two flavors of LLVM in parallel. All the other tools can be
downloaded and built while we are waiting for LLVM.