[TRAFODION-1700] Upsert with omitted default value columns leave the aligned format table in corrupted state. [TRAFODION-1847] Upsert with omitted timestamp columns having current_timestamp as default in a non-aligned format table returns wrong value for this column
For TRAFODION-1847, the upsert is transformed into merge.
For TRAFODION-1700, the upsert is transformed into merge when the CQD
TRAF_UPSERT_WITH_INSERT_DEFAULT_SEMANTICS is set to OFF. By default this CQD is
set to OFF. When this CQD is ON, the upsert will add a new row with the omitted
columns filled with default values always.