[TRAFODION-1836] Remove win-odbc license info from clients package
Removing TFODBC64 from the client packaging, since there is still
a license issue with pulling in redistributed code from visual
studio.  The existing license info is moved to a new file that is
not added to the LICENSE file during packaging. But it is retained
for future use when the issue can be resolved and win-odbc added back
into the binary distribution.